+++ apps/shopify/app/lib/shopify/queries.server.js
@@ -54,6 +54,9 @@ export const PRODUCTS_QUERY = `#graphql
                 updatedAt
                 inventoryItem {
                   id
+                  unitCost {
+                    amount
+                  }
                 }
               }
             }